+++ src/tasks/build-app.js
@@ -1,5 +1,5 @@
 /* eslint camelcase: ["error", {properties: "never"}] */
-/*eslint-env node, mocha */
+/* eslint-env node, mocha */
 
 const _ = require('lodash');
 const path = require('path');
@@ -13,7 +13,6 @@ const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 
 module.exports = function(config, callback) {
-
     const wpConfig = config.webpackConfig.application;
     // use production optimizations
 
@@ -29,9 +28,9 @@ module.exports = function(config, callback) {
         // (http://ianobermiller.com/blog/2015/06/15/shave-45kb-off-your-production-webpack-react-build/)
         new webpack.DefinePlugin({
             'process.env': {NODE_ENV: '"production"'},
-            '__DEBUG__': false
+            __DEBUG__: false,
         }),
-        new webpack.ContextReplacementPlugin(/moment[\/\\]locale$/, /~$/),
+        new webpack.ContextReplacementPlugin(/moment[/\\]locale$/, /~$/),
         new ExtractTextPlugin({
             filename: 'style.css?[contenthash:5]',
             allChunks: true,
@@ -70,10 +69,10 @@ module.exports = function(config, callback) {
                 cssProcessorOptions: {
                     autoprefixer: {add: true, browsers: compatibleBrowsers},
                     discardComments: {
-                        removeAll: true
-                    }
+                        removeAll: true,
+                    },
                 },
-                canPrint: process.env.NODE_ENV !== 'test'
+                canPrint: process.env.NODE_ENV !== 'test',
             })
         );
     }
@@ -81,12 +80,10 @@ module.exports = function(config, callback) {
     optimizations.push(new webpack.optimize.UglifyJsPlugin(uglifyOptions));
 
     if (wpConfig.html) {
-
         const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-        //@deprecated
+        // @deprecated
         if (wpConfig.html.template && /\.html$/.test(wpConfig.html.template)) {
-
             gutil.log(
                 gutil.colors.yellow('[DEPRECATION]'),
                 'Please provide an .ejs template for html-webpack plugin and no .html template'
@@ -98,15 +95,12 @@ module.exports = function(config, callback) {
         }
 
         optimizations.push(new HtmlWebpackPlugin(wpConfig.html));
-
     }
 
     if (wpConfig.copyFiles) {
-
         const CopyWebpackPlugin = require('copy-webpack-plugin');
 
         optimizations.push(new CopyWebpackPlugin(wpConfig.copyFiles));
-
     }
 
     if (wpConfig.plugins) {