+++ src/tasks/test.js
@@ -2,9 +2,9 @@ const gulp = require('gulp');
 const mocha = require('gulp-spawn-mocha');
 
 module.exports = function(config) {
-    return gulp
-        .src(config.testEntryPoint)
-        .pipe(mocha({
+    return gulp.src(config.testEntryPoint).pipe(
+        mocha({
             compilers: 'jsx?:babel-core/register',
-        }));
+        })
+    );
 };