+++ src/util/doctor.js
@@ -4,11 +4,11 @@ const _ = require('lodash');
 const fs = require('fs-extra');
 const semver = require('semver');
 const cp = require('child_process');
+
 const execSync = cp.execSync;
 
 class Doctor {
     constructor(dir, config) {
-
         this.basedir = dir;
         this.pjsonFile = path.join(this.basedir, 'package.json');
         this.config = config || {};
@@ -17,14 +17,13 @@ class Doctor {
     }
 
     heal() {
-
         if (!this.hasFixableProblems()) {
             return 'No fixable problems found';
         }
 
         let messages = 'The doctor tries to heal...';
 
-        _.forEach(this.deleteCandidates, function(file) {
+        _.forEach(this.deleteCandidates, file => {
             messages += `\nTrying to delete ${file} ... `;
             try {
                 fs.removeSync(file);
@@ -49,14 +48,17 @@ class Doctor {
         this.check();
 
         return messages;
-
     }
 
     static asyncSelfCheck({dir, logger = console.log, callback = _.noop}) {
-
         const checkPackages = {
-            'ecc-dotfiles': path.join(dir, 'node_modules', 'ecc-dotfiles', 'package.json'),
-            'ecc-gulp-tasks': path.join(__dirname, '../../package.json')
+            'ecc-dotfiles': path.join(
+                dir,
+                'node_modules',
+                'ecc-dotfiles',
+                'package.json'
+            ),
+            'ecc-gulp-tasks': path.join(__dirname, '../../package.json'),
         };
 
         const msg = [];
@@ -65,9 +67,13 @@ class Doctor {
 
         const cb = () => {
             count -= 1;
-            if(count <= 0){
-                logger(`Version checks done. ${_.size(msg)} Dependencies out of date.`);
-                _.forEach(msg, (m) => logger(m));
+            if (count <= 0) {
+                logger(
+                    `Version checks done. ${_.size(
+                        msg
+                    )} Dependencies out of date.`
+                );
+                _.forEach(msg, m => logger(m));
                 callback();
             }
         };
@@ -76,27 +82,31 @@ class Doctor {
             logger(`Checking if there is a new version of ${dep}...`);
             try {
                 const dfVersion = fs.readJsonSync(pjson).version;
-                cp.exec(`yarn info ${dep} version --json`, function(err, stdout, stderr) {
-
-                    if (err || !_.isEmpty(stderr.toString())) {
-                        return;
-                    }
-
-                    try {
-                        const latest = JSON.parse(stdout.toString()).data;
-                        if (!semver.satisfies(dfVersion, latest)) {
-                            msg.push(`${dep}@${dfVersion} installed. Newest version is ${latest}. Please run \`yarn upgrade ${dep}\``);
+                cp.exec(
+                    `yarn info ${dep} version --json`,
+                    (err, stdout, stderr) => {
+                        if (err || !_.isEmpty(stderr.toString())) {
+                            return;
                         }
-                        cb();
-                    } catch (e) {
-                        cb();
+
+                        try {
+                            const latest = JSON.parse(stdout.toString()).data;
+                            if (!semver.satisfies(dfVersion, latest)) {
+                                msg.push(
+                                    `${dep}@${dfVersion} installed. Newest version is ${latest}. Please run \`yarn upgrade ${dep}\``
+                                );
+                            }
+                            cb();
+                        } catch (e) {
+                            cb();
+                        }
                     }
-                });
+                );
             } catch (e) {
                 cb();
             }
         });
-    };
+    }
 
     check() {
         this.messages = {};
@@ -122,125 +132,118 @@ class Doctor {
         this.checkGulpConfig();
 
         this.checkEnv();
-
     }
 
     checkEnv() {
         const messages = [];
 
-        const checkCommands = fs.readJsonSync(path.join(__dirname, '../../env.json'));
+        const checkCommands = fs.readJsonSync(
+            path.join(__dirname, '../../env.json')
+        );
 
-        _.forEach(checkCommands, ({cmd, version,}, tool) => {
-
+        _.forEach(checkCommands, ({cmd, version}, tool) => {
             try {
-                const installedVersion = execSync(
-                    `${tool} ${cmd}`,
-                    {
-                        cwd: this.basedir,
-                    })
+                const installedVersion = execSync(`${tool} ${cmd}`, {
+                    cwd: this.basedir,
+                })
                     .toString()
                     .replace(/\r?\n/g, '');
 
                 if (!semver.satisfies(installedVersion, `${version}`, true)) {
-
-                    let m = `You are using ${tool}@${installedVersion}.`
+                    let m = `You are using ${tool}@${installedVersion}.`;
                     m += ` The current recommended version is ${version}.`;
 
                     messages.push(m);
-
                 }
-
             } catch (e) {
                 // we die gracefully, as the check errored, or something
             }
-
         });
 
         if (!_.isEmpty(messages)) {
-            let envMessages = 'The following problems have been found with your environment:';
-            _.map(messages, (message) => {
+            let envMessages =
+                'The following problems have been found with your environment:';
+            _.map(messages, message => {
                 envMessages += `\n\t${message}`;
             });
 
-            const version = require(path.join(__dirname, '../../package.json')).version;
+            // eslint-disable-next-line
+            const version = require(path.join(__dirname, '../../package.json'))
+                .version;
 
-            const url = 'https://github.com/elds/ecc-gulp-tasks/blob/v' + version + '/README.md#environment';
+            const url = `https://github.com/elds/ecc-gulp-tasks/blob/v${version}/README.md#environment`;
 
-            envMessages += '\n\n\tPlease refer to ' + url + ' for more information/resolutions.';
+            envMessages += `\n\n\tPlease refer to ${url} for more information/resolutions.`;
 
             this.messages.envMessages = envMessages;
-
         }
-
-
     }
 
     checkGulpConfig() {
         const messages = [];
 
-        const deprectatedValues = [
-            'path',
-            'serverStart',
-            'serverOverrides',
-        ];
+        const deprectatedValues = ['path', 'serverStart', 'serverOverrides'];
 
-        _.forEach(deprectatedValues, (key) => {
+        _.forEach(deprectatedValues, key => {
             const value = _.get(this.config, key, false);
 
             if (value) {
-                messages.push(`Please delete deprecated option '${key}': '${value}' from your gulp buildConfig.`);
+                messages.push(
+                    `Please delete deprecated option '${key}': '${value}' from your gulp buildConfig.`
+                );
             }
         });
 
         if (!_.isEmpty(messages)) {
-            let gulpConfigMessage = 'The following problems are in this projects buildConfig:';
-            _.map(messages, (message) => {
+            let gulpConfigMessage =
+                'The following problems are in this projects buildConfig:';
+            _.map(messages, message => {
                 gulpConfigMessage += `\n\t${message}`;
             });
 
             this.messages.gulpConfig = gulpConfigMessage;
-
         }
-
-
     }
 
     checkPackageJson() {
-
         try {
             const originalPJSON = fs.readJSONSync(this.pjsonFile);
 
             const fixedPJSON = _.clone(originalPJSON);
 
-            const messages = [];
+            let messages = [];
 
-            const reactPeer = _.get(originalPJSON, ['peerDependencies', 'react']);
+
+            const reactPeer = _.get(originalPJSON, [
+                'peerDependencies',
+                'react',
+            ]);
             if (_.isString(reactPeer) && reactPeer !== '*') {
-
                 _.set(fixedPJSON, ['peerDependencies', 'react'], '*');
 
-                messages.push('React peer dependency should be \'*\' and not \'{reactPeer}\' (auto-fixable)');
+                messages.push(
+                    "React peer dependency should be '*' and not '{reactPeer}' (auto-fixable)"
+                );
             }
 
             if (!_.isEmpty(messages)) {
                 this.fixedPJSON = fixedPJSON;
 
-                let PJSONMessage = 'The following problems are in this projects package.json:';
-                _.map(messages, (message) => {
+                let PJSONMessage =
+                    'The following problems are in this projects package.json:';
+                _.map(messages, message => {
                     PJSONMessage += `\n\t${message}`;
                 });
 
                 this.messages.pjson = PJSONMessage;
             }
-
         } catch (e) {
             // no package.json found / unreadable
+            console.warn(e);
         }
-
     }
 
     checkNotExists(patterns) {
-
         this.deleteCandidates = _.union(
             this.deleteCandidates,
             globby.sync(patterns, {
@@ -250,15 +253,14 @@ class Doctor {
         );
 
         if (!_.isEmpty(this.deleteCandidates)) {
-            let deleteCandidates = 'The following files should be deleted (auto-fixable):';
-            _.map(this.deleteCandidates, (filePath) => {
+            let deleteCandidates =
+                'The following files should be deleted (auto-fixable):';
+            _.map(this.deleteCandidates, filePath => {
                 deleteCandidates += `\n\t${filePath}`;
             });
 
             this.messages.deleteCandidates = deleteCandidates;
-
         }
-
     }
 
     hasFixableProblems() {
@@ -270,24 +272,23 @@ class Doctor {
     }
 
     toString() {
-
         if (!this.hasProblems()) {
             return 'The doctor found no problems';
         }
 
         let result = 'The doctor found a few symptoms:';
 
-        _.map(this.messages, function(message) {
+        _.map(this.messages, message => {
             result += `\n\n${message}`;
         });
 
         if (this.hasFixableProblems()) {
-            result += '\n\nSome (or all) of these problems can be fixed automatically.';
+            result +=
+                '\n\nSome (or all) of these problems can be fixed automatically.';
             result += '\nPlease run `gulp doctor --heal` to fix them.';
         }
 
         return result;
-
     }
 }