+++ src/webpack/plugins/SCSSBannerPlugin.js
@@ -3,10 +3,10 @@ const path = require('path');
 const ConcatSource = require('webpack-sources').ConcatSource;
 
 function SCSSBannerPlugin(outputPath, outputFileName, styleSCSS) {
-    this.banner = '// This is necessary so that components may use variables in scss\n' +
-        'if(__WEBPACK__){\n' +
-        '  require(\'' + path.relative(outputPath, styleSCSS) + '\');\n' +
-        '}';
+    this.banner = `// This is necessary so that components may use variables in scss
+if(__WEBPACK__){
+  require('${path.relative(outputPath, styleSCSS)}');
+}`;
     this.filename = path.basename(outputFileName);
 }
 
@@ -16,19 +16,19 @@ SCSSBannerPlugin.prototype.apply = function(compiler) {
     const banner = this.banner;
     const filename = this.filename;
 
-    compiler.plugin('compilation', function(compilation) {
-        compilation.plugin('optimize-chunk-assets', function(chunks, callback) {
-            chunks.forEach(function(chunk) {
-                chunk.files
-                    .filter(function(file) {
-                        return file === filename;
-                    })
-                    .forEach(function(file) {
-                        compilation.assets[file] = new ConcatSource(banner, '\n', compilation.assets[file]);
-                    });
+    compiler.plugin('compilation', compilation => {
+        compilation.plugin('optimize-chunk-assets', (chunks, callback) => {
+            chunks.forEach(chunk => {
+                chunk.files.filter(file => file === filename).forEach(file => {
+                    // eslint-disable-next-line no-param-reassign
+                    compilation.assets[file] = new ConcatSource(
+                        banner,
+                        '\n',
+                        compilation.assets[file]
+                    );
+                });
             });
             callback();
         });
-
     });
 };