+++ src/webpack/utils.js
@@ -5,41 +5,46 @@ function mergeFunction(objValue, srcValue) {
     if (_.isArray(objValue)) {
         return _.concat(objValue, srcValue);
     }
+    return undefined;
 }
 
-const isJS = (userRequest) => {
-    return _.endsWith(userRequest, '.js') || _.endsWith(userRequest, '.jsx');
-}
+const isJS = userRequest =>
+    _.endsWith(userRequest, '.js') || _.endsWith(userRequest, '.jsx');
 
-const isEccenca = (module) => {
+const isEccenca = module => {
     let userRequest = module.userRequest;
 
-    if (!_.isString(userRequest) || !(isJS(userRequest))) {
+    if (!_.isString(userRequest) || !isJS(userRequest)) {
         return false;
     }
 
     userRequest = slash(userRequest);
 
-    return userRequest.lastIndexOf('node_modules/ecc-') >= 0 &&
-        userRequest.lastIndexOf('node_modules/ecc-') === userRequest.lastIndexOf('node_modules');
+    return (
+        userRequest.lastIndexOf('node_modules/ecc-') >= 0 &&
+        userRequest.lastIndexOf('node_modules/ecc-') ===
+            userRequest.lastIndexOf('node_modules')
+    );
 };
 
-const isExternal = (module) => {
+const isExternal = module => {
     let userRequest = module.userRequest;
 
-    if (!_.isString(userRequest) || !(isJS(userRequest)) || isEccenca(module)) {
+    if (!_.isString(userRequest) || !isJS(userRequest) || isEccenca(module)) {
         return false;
     }
 
     userRequest = slash(userRequest);
 
-    return userRequest.indexOf('bower_components') >= 0 ||
+    return (
+        userRequest.indexOf('bower_components') >= 0 ||
         userRequest.indexOf('node_modules') >= 0 ||
-        userRequest.indexOf('libraries') >= 0;
+        userRequest.indexOf('libraries') >= 0
+    );
 };
 
 module.exports = {
     mergeFunction,
     isEccenca,
-    isExternal
+    isExternal,
 };