+++ test/lint.test.js
@@ -1,34 +1,27 @@
-var should = require('should');
-var lint = require('../src/tasks/lint');
+const should = require('should');
+const lint = require('../src/tasks/lint');
 
-describe('eslint', function() {
-
-    it('should fail with an invalid js file', function(done) {
-        var stream = lint({
-            lintingFiles: [
-                './test/fixtures/lint/fail.js'
-            ]
+describe('eslint', () => {
+    it('should fail with an invalid js file', done => {
+        const stream = lint({
+            lintingFiles: ['./test/fixtures/lint/fail.js'],
         });
 
-        stream.on('error', function(err) {
-            (function() {
-                throw err
-            }).should.throw(/Unexpected var/);
+        stream.on('error', err => {
+            should(() => {
+                throw err;
+            }).throw(/Unexpected var/);
             done();
-        })
+        });
     });
 
-    it('should not fail with a valid js file', function(done) {
-        var stream = lint({
-            lintingFiles: [
-                './test/fixtures/lint/good.js'
-            ]
+    it('should not fail with a valid js file', done => {
+        const stream = lint({
+            lintingFiles: ['./test/fixtures/lint/good.js'],
         });
 
         stream.on('end', done);
 
-        stream.on('data', function(){});
-
+        stream.on('data', () => {});
     });
-
 });